Move gtf annotation to gtf.py to keep it consistent
